requirements.txt
pandas==2.1.4
openpyxl==3.1.2
sentence-transformers==2.2.2
pymilvus==2.3.4
requests==2.31.0
python-docx==1.1.0
python-dotenv==1.0.0
pydantic==2.5.2
numpy==1.24.3
scikit-learn==1.3.2 